+++ packages/clients/src/api/container/v1beta1/marshalling.gen.ts
@@ -10,6 +10,7 @@ import {
 import type { DefaultValues } from '../../../bridge'
 import type {
   Container,
+  ContainerScalingOption,
   CreateContainerRequest,
   CreateCronRequest,
   CreateDomainRequest,
@@ -42,6 +43,20 @@ import type {
   UpdateTriggerRequestSqsClientConfig,
 } from './types.gen'
 
+const unmarshalContainerScalingOption = (
+  data: unknown,
+): ContainerScalingOption => {
+  if (!isJSONObject(data)) {
+    throw new TypeError(
+      `Unmarshalling the type 'ContainerScalingOption' failed as data isn't a dictionary.`,
+    )
+  }
+
+  return {
+    concurrentRequestsThreshold: data.concurrent_requests_threshold,
+  } as ContainerScalingOption
+}
+
 const unmarshalSecretHashedValue = (data: unknown): SecretHashedValue => {
   if (!isJSONObject(data)) {
     throw new TypeError(
@@ -64,6 +79,7 @@ export const unmarshalContainer = (data: unknown): Container => {
 
   return {
     cpuLimit: data.cpu_limit,
+    createdAt: unmarshalDate(data.created_at),
     description: data.description,
     domainName: data.domain_name,
     environmentVariables: data.environment_variables,
@@ -80,15 +96,20 @@ export const unmarshalContainer = (data: unknown): Container => {
     port: data.port,
     privacy: data.privacy,
     protocol: data.protocol,
+    readyAt: unmarshalDate(data.ready_at),
     region: data.region,
     registryImage: data.registry_image,
     sandbox: data.sandbox,
+    scalingOption: data.scaling_option
+      ? unmarshalContainerScalingOption(data.scaling_option)
+      : undefined,
     secretEnvironmentVariables: unmarshalArrayOfObject(
       data.secret_environment_variables,
       unmarshalSecretHashedValue,
     ),
     status: data.status,
     timeout: data.timeout,
+    updatedAt: unmarshalDate(data.updated_at),
   } as Container
 }
 
@@ -340,6 +361,18 @@ export const unmarshalListTriggersResponse = (
   } as ListTriggersResponse
 }
 
+const marshalContainerScalingOption = (
+  request: ContainerScalingOption,
+  defaults: DefaultValues,
+): Record<string, unknown> => ({
+  ...resolveOneOf([
+    {
+      param: 'concurrent_requests_threshold',
+      value: request.concurrentRequestsThreshold,
+    },
+  ]),
+})
+
 const marshalSecret = (
   request: Secret,
   defaults: DefaultValues,
@@ -368,6 +401,10 @@ export const marshalCreateContainerRequest = (
   protocol: request.protocol,
   registry_image: request.registryImage,
   sandbox: request.sandbox,
+  scaling_option:
+    request.scalingOption !== undefined
+      ? marshalContainerScalingOption(request.scalingOption, defaults)
+      : undefined,
   secret_environment_variables:
     request.secretEnvironmentVariables !== undefined
       ? request.secretEnvironmentVariables.map(elt =>
@@ -512,6 +549,10 @@ export const marshalUpdateContainerRequest = (
   redeploy: request.redeploy,
   registry_image: request.registryImage,
   sandbox: request.sandbox,
+  scaling_option:
+    request.scalingOption !== undefined
+      ? marshalContainerScalingOption(request.scalingOption, defaults)
+      : undefined,
   secret_environment_variables:
     request.secretEnvironmentVariables !== undefined
       ? request.secretEnvironmentVariables.map(elt =>