+++ packages/clients/src/api/iam/v1alpha1/index.gen.ts
@@ -86,6 +86,7 @@ export type {
   UpdateGroupRequest,
   UpdatePolicyRequest,
   UpdateSSHKeyRequest,
+  UpdateUserPasswordRequest,
   UpdateUserRequest,
   User,
   UserStatus,