+++ packages/clients/src/api/k8s/v1/types.gen.ts
@@ -1140,16 +1140,6 @@ export interface ListVersionsResponse {
   versions: Version[]
 }
 
-export type MigrateClusterToRoutedIPsRequest = {
-  /**
-   * Region to target. If none is passed will use default region from the
-   * config.
-   */
-  region?: Region
-  /** Cluster ID for which the routed ip will be enabled for the nodes. */
-  clusterId: string
-}
-
 export type MigrateClusterToSBSCSIRequest = {
   /**
    * Region to target. If none is passed will use default region from the